app/requirements.txt
Flask~=2.2.2
Flask_HTTPAuth~=4.7.0
Werkzeug~=2.2.2
PyYAML~=6.0
Requests~=2.28.1